uses built-in emacs 26 --fg-daemon flag to de-dup my buggy wrappers
I had written a couple ruby programs that I'd wanted to keep an eye on
the emacs process, but in reality they just let it crash without
detecting it. Now that emacs has the `--fg-daemon` built in, we can
achieve basically the same thing.

From now on, `EDITOR` should be set to `emacsclient -t` and `em`
should be defined as the same in the shell config, rather than
referencing `~/bin`.