upgrades the emacs systemd service to use `--fg-daemon`
This is simpler because now systemd can send kill signals to the
non-forking emacs process, rather than having to start up a whole new
command string to connect to the daemon and ask it to quit.

IIRC this option was added in Emacs 26, which all of my machines are
now using.